trust changeset moves into trust module
