+++ crates/pyo3/src/lib.rs
@@ -10,7 +10,6 @@ use pyo3::prelude::*;
 
 pub mod acl;
 pub mod analysis;
-pub mod change;
 pub mod daemon;
 pub mod rules;
 pub mod system;
@@ -20,7 +19,6 @@ pub mod trust;
 fn rust(_py: Python, m: &PyModule) -> PyResult<()> {
     acl::init_module(_py, m)?;
     analysis::init_module(_py, m)?;
-    change::init_module(_py, m)?;
     daemon::init_module(_py, m)?;
     rules::init_module(_py, m)?;
     system::init_module(_py, m)?;
