+++ crates/pyo3/src/trust.rs
@@ -6,9 +6,12 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+use std::collections::HashMap;
+
 use pyo3::prelude::*;
 
 use fapolicy_api::trust::Trust;
+use fapolicy_trust::ops::{get_path_action_map, Changeset};
 use fapolicy_trust::stat::{Actual, Status};
 
 /// Trust entry
@@ -102,7 +105,61 @@ impl PyActual {
     }
 }
 
+/// A mutable collection of changes
+#[pyclass(module = "trust", name = "Changeset")]
+#[derive(Clone)]
+pub struct PyChangeset {
+    rs: Changeset,
+}
+
+impl From<Changeset> for PyChangeset {
+    fn from(rs: Changeset) -> Self {
+        Self { rs }
+    }
+}
+
+impl From<PyChangeset> for Changeset {
+    fn from(py: PyChangeset) -> Self {
+        py.rs
+    }
+}
+
+impl Default for PyChangeset {
+    fn default() -> Self {
+        Changeset::new().into()
+    }
+}
+
+#[pymethods]
+impl PyChangeset {
+    #[new]
+    pub fn new() -> Self {
+        PyChangeset::default()
+    }
+
+    pub fn add_trust(&mut self, path: &str) {
+        self.rs.add(path)
+    }
+
+    pub fn del_trust(&mut self, path: &str) {
+        self.rs.del(path)
+    }
+
+    pub fn len(&self) -> usize {
+        self.rs.len()
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    pub fn get_path_action_map(&self) -> HashMap<String, String> {
+        get_path_action_map(&self.rs)
+    }
+}
+
 pub fn init_module(_py: Python, m: &PyModule) -> PyResult<()> {
+    m.add_class::<PyChangeset>()?;
     m.add_class::<PyTrust>()?;
     m.add_class::<PyActual>()?;
     Ok(())
